// movie-store-web-app/src/app/api/api-reference.ts
export abstract class Movie implements IMovie {
    id?: string;
    name!: string;
    licensingType!: LicensingType;

    constructor(data?: IMovie) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.name = _data["name"];
            this.licensingType = _data["licensingType"];
        }
    }

    static fromJS(data: any): Movie {
        data = typeof data === 'object' ? data : {};
        throw new Error("The abstract class 'Movie' cannot be instantiated.");
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["name"] = this.name;
        data["licensingType"] = this.licensingType;
        return data;
    }
}
// ...
export class Customer implements ICustomer {
    customerStatus?: CustomerStatus;
    id?: string;
    email?: Email;
    purchasedMovies?: PurchasedMovie[];
    role!: CustomerRole;
    moneySpent?: Money;

    constructor(data?: ICustomer) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.customerStatus = _data["customerStatus"] ? CustomerStatus.fromJS(_data["customerStatus"]) : <any>undefined;
            this.id = _data["id"];
            this.email = _data["email"] ? Email.fromJS(_data["email"]) : <any>undefined;
            if (Array.isArray(_data["purchasedMovies"])) {
                this.purchasedMovies = [] as any;
                for (let item of _data["purchasedMovies"])
                    this.purchasedMovies!.push(PurchasedMovie.fromJS(item));
            }
            this.role = _data["role"];
            this.moneySpent = _data["moneySpent"] ? Money.fromJS(_data["moneySpent"]) : <any>undefined;
        }
    }

    static fromJS(data: any): Customer {
        data = typeof data === 'object' ? data : {};
        let result = new Customer();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["customerStatus"] = this.customerStatus ? this.customerStatus.toJSON() : <any>undefined;
        data["id"] = this.id;
        data["email"] = this.email ? this.email.toJSON() : <any>undefined;
        if (Array.isArray(this.purchasedMovies)) {
            data["purchasedMovies"] = [];
            for (let item of this.purchasedMovies)
                data["purchasedMovies"].push(item.toJSON());
        }
        data["role"] = this.role;
        data["moneySpent"] = this.moneySpent ? this.moneySpent.toJSON() : <any>undefined;
        return data;
    }
}

export interface ICustomer {
    customerStatus?: CustomerStatus;
    id?: string;
    email?: Email;
    purchasedMovies?: PurchasedMovie[];
    role: CustomerRole;
    moneySpent?: Money;
}

export class CustomerStatus implements ICustomerStatus {
    status?: Status;
    statusExpirationDate?: ExpirationDate;
    isAdvanced?: boolean;

    constructor(data?: ICustomerStatus) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.status = _data["status"];
            this.statusExpirationDate = _data["statusExpirationDate"] ? ExpirationDate.fromJS(_data["statusExpirationDate"]) : <any>undefined;
            this.isAdvanced = _data["isAdvanced"];
        }
    }

    static fromJS(data: any): CustomerStatus {
        data = typeof data === 'object' ? data : {};
        let result = new CustomerStatus();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["status"] = this.status;
        data["statusExpirationDate"] = this.statusExpirationDate ? this.statusExpirationDate.toJSON() : <any>undefined;
        data["isAdvanced"] = this.isAdvanced;
        return data;
    }
}

export interface ICustomerStatus {
    status?: Status;
    statusExpirationDate?: ExpirationDate;
    isAdvanced?: boolean;
}

export enum Status {
    Regular = 1,
    Advanced = 2,
}

export class ExpirationDate implements IExpirationDate {
    value?: Date | undefined;
    isExpired?: boolean;

    constructor(data?: IExpirationDate) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.value = _data["value"] ? new Date(_data["value"].toString()) : <any>undefined;
            this.isExpired = _data["isExpired"];
        }
    }

    static fromJS(data: any): ExpirationDate {
        data = typeof data === 'object' ? data : {};
        let result = new ExpirationDate();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["value"] = this.value ? this.value.toISOString() : <any>undefined;
        data["isExpired"] = this.isExpired;
        return data;
    }
}

export interface IExpirationDate {
    value?: Date | undefined;
    isExpired?: boolean;
}

export class Email implements IEmail {
    value?: string;

    constructor(data?: IEmail) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.value = _data["value"];
        }
    }

    static fromJS(data: any): Email {
        data = typeof data === 'object' ? data : {};
        let result = new Email();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["value"] = this.value;
        return data;
    }
}

export interface IEmail {
    value?: string;
}

export class PurchasedMovie implements IPurchasedMovie {
    id?: string;
    movie?: Movie;
    purchaseDate!: Date;
    expirationDate?: ExpirationDate;
    customer?: Customer;

    constructor(data?: IPurchasedMovie) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.movie = _data["movie"] ? Movie.fromJS(_data["movie"]) : <any>undefined;
            this.purchaseDate = _data["purchaseDate"] ? new Date(_data["purchaseDate"].toString()) : <any>undefined;
            this.expirationDate = _data["expirationDate"] ? ExpirationDate.fromJS(_data["expirationDate"]) : <any>undefined;
            this.customer = _data["customer"] ? Customer.fromJS(_data["customer"]) : <any>undefined;
        }
    }

    static fromJS(data: any): PurchasedMovie {
        data = typeof data === 'object' ? data : {};
        let result = new PurchasedMovie();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["movie"] = this.movie ? this.movie.toJSON() : <any>undefined;
        data["purchaseDate"] = this.purchaseDate ? this.purchaseDate.toISOString() : <any>undefined;
        data["expirationDate"] = this.expirationDate ? this.expirationDate.toJSON() : <any>undefined;
        data["customer"] = this.customer ? this.customer.toJSON() : <any>undefined;
        return data;
    }
}

export interface IPurchasedMovie {
    id?: string;
    movie?: Movie;
    purchaseDate: Date;
    expirationDate?: ExpirationDate;
    customer?: Customer;
}

export enum CustomerRole {
    Regular = 1,
    Admin = 2,
}

export class Money implements IMoney {
    value?: number;

    constructor(data?: IMoney) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.value = _data["value"];
        }
    }

    static fromJS(data: any): Money {
        data = typeof data === 'object' ? data : {};
        let result = new Money();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["value"] = this.value;
        return data;
    }
}

export interface IMoney {
    value?: number;
}

export class UpdateCustomerCommand implements IUpdateCustomerCommand {
    customerId?: string;
    email?: string;
    role?: CustomerRole;

    constructor(data?: IUpdateCustomerCommand) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.customerId = _data["customerId"];
            this.email = _data["email"];
            this.role = _data["role"];
        }
    }

    static fromJS(data: any): UpdateCustomerCommand {
        data = typeof data === 'object' ? data : {};
        let result = new UpdateCustomerCommand();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["customerId"] = this.customerId;
        data["email"] = this.email;
        data["role"] = this.role;
        return data;
    }
}

export interface IUpdateCustomerCommand {
    customerId?: string;
    email?: string;
    role?: CustomerRole;
}